requirements.txt
alembic==1.0.5
asn1crypto==1.2.0
atomicwrites==1.3.0
attrs==19.3.0
Automat==0.7.0
awscli==1.13.0
banal==0.4.0
bcrypt==3.1.7
bleach==3.1.0
boto==2.49.0
boto3==1.9.70
botocore==1.12.70
bz2file==0.98
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
Click==7.0
colorama==0.3.7
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.2
Cython==0.29.2
cytoolz==0.9.0.1
dataset==1.1.0
dill==0.2.8.2
docutils==0.14
en-core-web-sm==2.0.0
Flask==1.0.2
gensim==3.6.0
get==2019.4.13
gunicorn==19.9.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.23
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.3
kiwisolver==1.0.1
lxml==4.4.1
Mako==1.0.7
MarkupSafe==1.1.0
matplotlib==3.0.2
more-itertools==7.2.0
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.1
mysqlclient==1.3.14
normality==0.6.1
numpy==1.15.4
packaging==19.2
pandas==0.23.4
paramiko==2.4.2
parsel==1.5.2
pip-review==1.0
pkginfo==1.5.0.1
plac==0.9.6
pluggy==0.13.0
post==2019.4.13
preshed==2.0.1
Protego==0.1.15
public==2019.4.13
py==1.8.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
PyDispatcher==2.0.5
Pygments==2.3.1
PyHamcrest==1.9.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.2
PySocks==1.7.1
pytest==5.2.2
pytest-runner==5.1
python-dateutil==2.7.5
python-editor==1.0.3
pytz==2018.7
PyYAML==3.13
query-string==2019.4.13
queuelib==1.5.0
readme-renderer==24.0
regex==2018.1.10
request==2019.4.13
requests==2.22.0
requests-toolbelt==0.9.1
rsa==3.4.2
s3transfer==0.1.13
scikit-learn==0.20.2
scipy==1.2.0
Scrapy==1.8.0
service-identity==18.1.0
six==1.12.0
smart-open==1.7.1
spacy==2.0.18
SQLAlchemy==1.2.15
thinc==6.12.1
toolz==0.9.0
tqdm==4.30.0
twine==1.12.1
Twisted==19.7.0
ujson==1.35
urllib3==1.24.2
w3lib==1.21.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
wrapt==1.10.11
xgboost==0.81
zipp==0.6.0
zope.interface==4.6.0
